# openapi/corvus-mvp.json
{
  "components": {
    "schemas": {
      "AccountProfile": {
        "additionalProperties": false,
        "properties": {
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "experience_kind": {
            "$ref": "#/components/schemas/ExperienceKind"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "entity_id",
          "experience_kind",
          "version"
        ],
        "title": "AccountProfile",
        "type": "object"
      },
      "AccountProfilePayload": {
        "additionalProperties": false,
        "properties": {
          "experience_kind": {
            "$ref": "#/components/schemas/ExperienceKind"
          }
        },
        "required": [
          "experience_kind"
        ],
        "title": "AccountProfilePayload",
        "type": "object"
      },
      "ApiErrorDetail": {
        "additionalProperties": true,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "format": "uuid",
            "title": "Correlation Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "correlation_id"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
      "ApiErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "ExperienceKind": {
        "enum": [
          "everyday",
          "developer"
        ],
        "title": "ExperienceKind",
        "type": "string"
      },
...
      "RecordStatus": {
        "enum": [
          "active",
          "suspended",
          "revoked",
          "archived"
        ],
        "title": "RecordStatus",
        "type": "string"
      },
...
      "SyncApplyResult": {
        "additionalProperties": false,
        "properties": {
          "acknowledged_cursor": {
            "minimum": 0.0,
            "title": "Acknowledged Cursor",
            "type": "integer"
          },
          "results": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SyncMutationResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "acknowledged_cursor"
        ],
        "title": "SyncApplyResult",
        "type": "object"
      },
      "SyncMutation": {
        "additionalProperties": false,
        "properties": {
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "kind": {
            "enum": [
              "account_profile",
              "workspace_profile"
            ],
            "title": "Kind",
            "type": "string"
          },
          "operation": {
            "enum": [
              "set_experience",
              "update"
            ],
            "title": "Operation",
            "type": "string"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountProfilePayload"
              },
              {
                "$ref": "#/components/schemas/WorkspaceProfilePayload"
              }
            ],
            "title": "Payload"
          }
        },
        "required": [
          "idempotency_key",
          "kind",
          "operation",
          "entity_id",
          "expected_version",
          "payload"
        ],
        "title": "SyncMutation",
        "type": "object"
      },
      "SyncMutationBatch": {
        "additionalProperties": false,
        "properties": {
          "acknowledged_cursor": {
            "minimum": 0.0,
            "title": "Acknowledged Cursor",
            "type": "integer"
          },
          "mutations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SyncMutation"
            },
            "maxItems": 100,
            "title": "Mutations",
            "type": "array"
          }
        },
        "required": [
          "acknowledged_cursor"
        ],
        "title": "SyncMutationBatch",
        "type": "object"
      },
      "SyncMutationResult": {
        "additionalProperties": false,
        "properties": {
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_version": {
            "minimum": 1.0,
            "title": "Entity Version",
            "type": "integer"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "kind": {
            "enum": [
              "account_profile",
              "workspace_profile"
            ],
            "title": "Kind",
            "type": "string"
          },
          "operation": {
            "enum": [
              "set_experience",
              "update"
            ],
            "title": "Operation",
            "type": "string"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountProfile"
              },
              {
                "$ref": "#/components/schemas/WorkspaceProfile"
              }
            ],
            "title": "Profile"
          },
          "sequence": {
            "minimum": 1.0,
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "idempotency_key",
          "kind",
          "operation",
          "entity_id",
          "entity_version",
          "sequence",
          "profile"
        ],
        "title": "SyncMutationResult",
        "type": "object"
      },
      "SyncPage": {
        "additionalProperties": false,
        "properties": {
          "changes": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/WorkspaceChange"
            },
            "title": "Changes",
            "type": "array"
          },
          "earliest_retained_sequence": {
            "minimum": 1.0,
            "title": "Earliest Retained Sequence",
            "type": "integer"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "high_watermark": {
            "minimum": 0.0,
            "title": "High Watermark",
            "type": "integer"
          },
          "next_cursor": {
            "minimum": 0.0,
            "title": "Next Cursor",
            "type": "integer"
          },
          "requested_cursor": {
            "minimum": 0.0,
            "title": "Requested Cursor",
            "type": "integer"
          }
        },
        "required": [
          "requested_cursor",
          "next_cursor",
          "high_watermark",
          "earliest_retained_sequence",
          "has_more"
        ],
        "title": "SyncPage",
        "type": "object"
      },
...
      "WorkspaceChange": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "change_digest": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Change Digest",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "device_version": {
            "minimum": 1.0,
            "title": "Device Version",
            "type": "integer"
          },
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_version": {
            "minimum": 1.0,
            "title": "Entity Version",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "account_profile",
              "workspace_profile"
            ],
            "title": "Kind",
            "type": "string"
          },
          "membership_version": {
            "minimum": 1.0,
            "title": "Membership Version",
            "type": "integer"
          },
          "operation": {
            "enum": [
              "set_experience",
              "update"
            ],
            "title": "Operation",
            "type": "string"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountProfile"
              },
              {
                "$ref": "#/components/schemas/WorkspaceProfile"
              }
            ],
            "title": "Payload"
          },
          "previous_digest": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Previous Digest",
            "type": "string"
          },
          "principal_id": {
            "format": "uuid",
            "title": "Principal Id",
            "type": "string"
          },
          "sequence": {
            "minimum": 1.0,
            "title": "Sequence",
            "type": "integer"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          },
          "workspace_version": {
            "minimum": 1.0,
            "title": "Workspace Version",
            "type": "integer"
          }
        },
        "required": [
          "workspace_id",
          "workspace_version",
          "sequence",
          "previous_digest",
          "change_digest",
          "kind",
          "operation",
          "entity_id",
          "entity_version",
          "payload",
          "account_id",
          "principal_id",
          "membership_version",
          "device_id",
          "device_version"
        ],
        "title": "WorkspaceChange",
        "type": "object"
      },
...
      "WorkspaceKind": {
        "enum": [
          "individual",
          "team"
        ],
        "title": "WorkspaceKind",
        "type": "string"
      },
      "WorkspaceProfile": {
        "additionalProperties": false,
        "properties": {
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RecordStatus"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          },
          "workspace_kind": {
            "$ref": "#/components/schemas/WorkspaceKind"
          }
        },
        "required": [
          "entity_id",
          "name",
          "workspace_kind",
          "status",
          "version"
        ],
        "title": "WorkspaceProfile",
        "type": "object"
      },
      "WorkspaceProfilePayload": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "workspace_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspaceKind"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "WorkspaceProfilePayload",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Corvus Hackathon MVP API",
    "version": "0.2.0-hackathon"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/workspaces/{workspace_id}/sync": {
      "get": {
        "operationId": "unavailable_workspace_sync_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authority denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Version or sync conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Payload rejected"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "summary": "Unavailable Workspace Sync Get",
        "tags": [
          "platform-sync"
        ]
      }
    },
    "/api/v2/workspaces/{workspace_id}/sync/mutations": {
      "post": {
        "operationId": "unavailable_workspace_sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncMutationBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncApplyResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authority denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Version or sync conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Payload rejected"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "summary": "Unavailable Workspace Sync Post",
        "tags": [
          "platform-sync"
        ]
      }
    },
...
  }
}